requirements.txt
playwright>=1.40.0
pytest>=7.4.0
pytest-playwright>=0.4.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
httpx>=0.25.0
allure-pytest>=2.13.0
